Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 int main()
 {
-    int a;
+    float a;
     printf("enter the value for a");
-    scanf("%d",&a);
+    scanf("%f",&a);
     printf("You entered: ");
     return 0;
